requirements.txt
scikit-learn
pandas
matplotlib
joblib
mlflow
pyyaml
seaborn